+++ src/KingmakerGunslinger/Deeds/TrueGritDeed.cs
@@ -1,0 +1,24 @@
+namespace KingmakerGunslinger.Deeds
+{
+    internal enum TrueGritDeed
+    {
+        Deadeye,
+        GunslingersDodge,
+        QuickClear,
+        GunslingerInitiative,
+        PistolWhip,
+        StopBleeding,
+        DeadShot,
+        StartlingShot,
+        TargetingHead,
+        TargetingTorso,
+        TargetingLegs,
+        BleedingWound,
+        ExpertLoading,
+        LightningReload,
+        Evasive,
+        MenacingShot,
+        CheatDeath,
+        StunningShot
+    }
+}